• transferfunctions.py   - rename LinearCombination as LinearTransform and move to combinationfunctions.py
• combinationfunctions.py
  - add LinearTransform

PASSES ALL TESTS
