+++ psyneulink/core/components/functions/nonstateful/combinationfunctions.py
@@ -15,6 +15,7 @@
 * `Reduce`
 * `LinearCombination`
 * `CombineMeans`
+* `LinearTransform`
 * `PredictionErrorDeltaFunction`
 
 Overview
@@ -32,8 +33,11 @@ when the CombinationFunction is used as the function of an InputPort or OutputPo
 """
 
 import numbers
+import types
+import warnings
 
 import numpy as np
+
 try:
     import torch
 except ImportError:
@@ -43,20 +47,28 @@ from beartype import beartype
 from psyneulink._typing import Optional, Union, Literal
 
 from psyneulink.core import llvm as pnlvm
-from psyneulink.core.components.functions.function import Function_Base, FunctionError, FunctionOutputType
+from psyneulink.core.components.functions import function
+from psyneulink.core.components.functions.function import (
+    Function_Base, FunctionError, FunctionOutputType, function_keywords, get_matrix)
+from psyneulink.core.components.shellclasses import Projection
 from psyneulink.core.globals.keywords import \
-    ADDITIVE_PARAM, ARRANGEMENT, COMBINATION_FUNCTION_TYPE, COMBINE_MEANS_FUNCTION, CONCATENATE_FUNCTION, \
-    CROSS_ENTROPY, DEFAULT_VARIABLE, EXPONENTS, LINEAR_COMBINATION_FUNCTION, MULTIPLICATIVE_PARAM, OFFSET, OPERATION, \
-    PREDICTION_ERROR_DELTA_FUNCTION, PRODUCT, REARRANGE_FUNCTION, REDUCE_FUNCTION, SCALE, SUM, WEIGHTS, \
-    PREFERENCE_SET_NAME
-from psyneulink.core.globals.utilities import convert_all_elements_to_np_array, convert_to_np_array, is_numeric, is_numeric_scalar, np_array_less_than_2d, ValidParamSpecType
+    (ADDITIVE_PARAM, ARRANGEMENT, COMBINATION_FUNCTION_TYPE, COMBINE_MEANS_FUNCTION, CONCATENATE_FUNCTION, \
+     CROSS_ENTROPY, DEFAULT_VARIABLE, EXPONENTS,
+     HAS_INITIALIZERS, HOLLOW_MATRIX, IDENTITY_MATRIX,
+     LINEAR_COMBINATION_FUNCTION, LINEAR_TRANSFORM_FUNCTION, \
+     MATRIX_KEYWORD_NAMES, MATRIX, MULTIPLICATIVE_PARAM, NORMALIZE,
+     OFFSET, OPERATION, PREDICTION_ERROR_DELTA_FUNCTION, PRODUCT,
+     REARRANGE_FUNCTION, RECEIVER, REDUCE_FUNCTION, SCALE, SUM, WEIGHTS, PREFERENCE_SET_NAME)
+from psyneulink.core.globals.utilities import (
+    convert_all_elements_to_np_array, convert_to_np_array, is_numeric, is_matrix_keyword, is_numeric_scalar,
+    np_array_less_than_2d, ValidParamSpecType)
 from psyneulink.core.globals.context import ContextFlags, handle_external_context
-from psyneulink.core.globals.parameters import Parameter, check_user_specified
+from psyneulink.core.globals.parameters import Parameter, check_user_specified, copy_parameter_value
 from psyneulink.core.globals.preferences.basepreferenceset import \
     REPORT_OUTPUT_PREF, ValidPrefSet, PreferenceEntry, PreferenceLevel
 
-__all__ = ['CombinationFunction', 'Concatenate', 'CombineMeans', 'Rearrange', 'Reduce', 'LinearCombination',
-           'PredictionErrorDeltaFunction']
+__all__ = ['CombinationFunction', 'Concatenate', 'CombineMeans', 'Rearrange', 'Reduce',
+           'LinearCombination', 'LinearTransform', 'PredictionErrorDeltaFunction']
 
 class CombinationFunction(Function_Base):
     """Function that combines multiple items, yielding a result with the same shape as its operands
@@ -1594,6 +1606,630 @@ class LinearCombination(
             from psyneulink.library.compositions.autodiffcomposition import AutodiffCompositionError
             raise AutodiffCompositionError(f"The 'operation' parameter of {function.componentName} is not supported "
                                            f"by AutodiffComposition; use 'SUM' or 'PRODUCT' if possible.")
+
+
+# **********************************************************************************************************************
+#                                                 LinearTransform
+# **********************************************************************************************************************
+
+class LinearTransform(CombinationFunction):  # -------------------------------------------------------------------------------
+    """
+    LinearTransform(          \
+         default_variable, \
+         matrix=None,      \
+         normalize=False,  \
+         params=None,      \
+         owner=None,       \
+         name=None,        \
+         prefs=None        \
+         )
+
+    .. _LinearTransform:
+
+    Matrix transform of `variable <LinearTransform.variable>`.
+
+    `function <LinearTransform._function>` returns dot product of variable with matrix:
+
+    .. math::
+        variable \\bullet matrix
+
+    If **normalize** is True, the result is normalized by the product of the norms of the variable and matrix:
+
+    .. math::
+        \\frac{variable \\bullet matrix}{\\|variable\\| \\cdot \\|matrix\\|}
+
+    COMMENT:  [CONVERT TO FIGURE]
+        ----------------------------------------------------------------------------------------------------------
+        MATRIX FORMAT <shape: (3,5)>
+                                         INDICES:
+                                     Output elements:
+                              0       1       2       3       4
+                         0  [0,0]   [0,1]   [0,2]   [0,3]   [0,4]
+        Input elements:  1  [1,0]   [1,1]   [1,2]   [1,3]   [1,4]
+                         2  [2,0]   [2,1]   [2,2]   [2,3]   [2,4]
+
+        matrix.shape => (input/rows, output/cols)
+
+        ----------------------------------------------------------------------------------------------------------
+        ARRAY FORMAT
+                                                                            INDICES
+                                          [ [      Input 0 (row0)       ], [       Input 1 (row1)      ]... ]
+                                          [ [ out0,  out1,  out2,  out3 ], [ out0,  out1,  out2,  out3 ]... ]
+        matrix[input/rows, output/cols]:  [ [ row0,  row0,  row0,  row0 ], [ row1,  row1,  row1,  row1 ]... ]
+                                          [ [ col0,  col1,  col2,  col3 ], [ col0,  col1,  col2,  col3 ]... ]
+                                          [ [[0,0], [0,1], [0,2], [0,3] ], [[1,0], [1,1], [1,2], [1,3] ]... ]
+
+        ----------------------------------------------------------------------------------------------------------
+    COMMENT
+
+
+    Arguments
+    ---------
+
+    variable : list or 1d array : default class_defaults.variable
+        specifies a template for the value to be transformed; length must equal the number of rows of `matrix
+        <LinearTransform.matrix>`.
+
+    matrix : number, list, 1d or 2d np.ndarray, function, or matrix keyword : default IDENTITY_MATRIX
+        specifies matrix used to transform `variable <LinearTransform.variable>`
+        (see `matrix <LinearTransform.matrix>` for specification details).
+
+        When LinearTransform is the `function <Projection_Base._function>` of a projection:
+
+            - the matrix specification must be compatible with the variables of the `sender <Projection_Base.sender>`
+              and `receiver <Projection_Base.receiver>`
+
+            - a matrix keyword specification generates a matrix based on the sender and receiver shapes
+
+        When LinearTransform is instantiated on its own, or as the function of a `Mechanism <Mechanism>` or `Port`:
+
+            - the matrix specification must be compatible with the function's own `variable <LinearTransform.variable>`
+
+            - if matrix is not specified, a square identity matrix is generated based on the number of columns in
+              `variable <LinearTransform.variable>`
+
+            - matrix keywords are not valid matrix specifications
+
+    normalize : bool : default False
+        specifies whether to normalize the result of `function <LinearCombination.function>` by dividing it by the
+        norm of `variable <LinearTransform.variable>` x the norm of `matrix <LinearTransform.matrix>`.
+
+    bounds : None
+
+    params : Dict[param keyword: param value] : default None
+        a `parameter dictionary <ParameterPort_Specification>` that specifies the parameters for the
+        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
+        arguments of the constructor.
+
+    owner : Component
+        `component <Component>` to which to assign the Function.
+
+    name : str : default see `name <Function.name>`
+        specifies the name of the Function.
+
+    prefs : PreferenceSet or specification dict : default Function.classPreferences
+        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
+
+    Attributes
+    ----------
+
+    variable : 1d array
+        contains value to be transformed.
+
+    matrix : 2d array
+        matrix used to transform `variable <LinearTransform.variable>`.
+        Can be specified as any of the following:
+            * number - used as the filler value for all elements of the :keyword:`matrix` (call to np.fill);
+            * list of arrays, 2d array - assigned as the value of :keyword:`matrix`;
+            * matrix keyword - see `MatrixKeywords` for list of options.
+        Rows correspond to elements of the input array (outer index), and
+        columns correspond to elements of the output array (inner index).
+
+    normalize : bool
+        determines whether the result of `function <LinearCombination.function>` is normalized, by dividing it by the
+        norm of `variable <LinearTransform.variable>` x the norm of `matrix <LinearTransform.matrix>`.
+
+
+    owner : Component
+        `component <Component>` to which the Function has been assigned.
+
+    name : str
+        the name of the Function; if it is not specified in the **name** argument of the constructor, a default is
+        assigned by FunctionRegistry (see `Registry_Naming` for conventions used for default and duplicate names).
+
+    prefs : PreferenceSet or specification dict : Function.classPreferences
+        the `PreferenceSet` for function; if it is not specified in the **prefs** argument of the Function's
+        constructor, a default is assigned using `classPreferences` defined in __init__.py (see `PreferenceSet`
+        for details).
+    """
+
+    componentName = LINEAR_TRANSFORM_FUNCTION
+
+    DEFAULT_FILLER_VALUE = 0
+
+    _model_spec_generic_type_name = 'onnx::MatMul'
+
+    class Parameters(CombinationFunction.Parameters):
+        """
+            Attributes
+            ----------
+
+                matrix
+                    see `matrix <LinearTransform.matrix>`
+
+                    :default value: None
+                    :type:
+
+                normalize
+                    see `normalize <LinearTransform.normalize>`
+
+                    :default value: False
+                    :type: bool
+        """
+        variable = Parameter(np.array([0]), read_only=True, pnl_internal=True, constructor_argument='default_variable', mdf_name='A')
+        matrix = Parameter(None, modulable=True, mdf_name='B')
+        normalize = Parameter(False)
+        bounds = None
+
+    # def is_matrix_spec(m):
+    #     if m is None:
+    #         return True
+    #     if m in MATRIX_KEYWORD_VALUES:
+    #         return True
+    #     if isinstance(m, (list, np.ndarray, types.FunctionType)):
+    #         return True
+    #     return False
+
+    @check_user_specified
+    @beartype
+    def __init__(self,
+                 default_variable=None,
+                 matrix=None,
+                 normalize=None,
+                 params=None,
+                 owner=None,
+                 prefs:  Optional[ValidPrefSet] = None):
+
+        # Note: this calls _validate_variable and _validate_params which are overridden below;
+        #       the latter implements the matrix if required
+        # super(LinearTransform, self).__init__(default_variable=default_variable,
+        super().__init__(
+            default_variable=default_variable,
+            matrix=matrix,
+            normalize=normalize,
+            params=params,
+            owner=owner,
+            prefs=prefs,
+        )
+
+        self.parameters.matrix.set(
+            self.instantiate_matrix(self.parameters.matrix.get()),
+            skip_log=True,
+        )
+
+    # def _validate_variable(self, variable, context=None):
+    #     """Insure that variable passed to LinearTransform is a max 2D array
+    #
+    #     :param variable: (max 2D array)
+    #     :param context:
+    #     :return:
+    #     """
+    #     variable = super()._validate_variable(variable, context)
+    #
+    #     # Check that variable <= 2D
+    #     try:
+    #         if not variable.ndim <= 2:
+    #             raise FunctionError("variable ({0}) for {1} must be a numpy.ndarray of dimension at most 2".format(variable, self.__class__.__name__))
+    #     except AttributeError:
+    #         raise FunctionError("PROGRAM ERROR: variable ({0}) for {1} should be a numpy.ndarray".
+    #                                 format(variable, self.__class__.__name__))
+    #
+    #     return variable
+
+
+    def _validate_params(self, request_set, target_set=None, context=None):
+        """Validate params and assign to targets
+
+        This overrides the class method, to perform more detailed type checking (see explanation in class method).
+        Note: this method (or the class version) is called only if the parameter_validation attribute is `True`
+
+        :param request_set: (dict) - params to be validated
+        :param target_set: (dict) - destination of validated params
+        :param context: (str)
+        :return none:
+        """
+
+        super()._validate_params(request_set, target_set, context)
+
+        param_set = target_set
+        # proxy for checking whether the owner is a projection
+        if hasattr(self.owner, "receiver"):
+            sender = self.defaults.variable
+            sender_len = np.size(np.atleast_2d(self.defaults.variable), 1)
+
+            # FIX: RELABEL sender -> input AND receiver -> output
+            # FIX: THIS NEEDS TO BE CLEANED UP:
+            #      - AT LEAST CHANGE THE NAME FROM kwReceiver TO output_template OR SOMETHING LIKE THAT
+            #      - MAKE ARG?  OR ADD OTHER PARAMS:  E.G., FILLER?
+            #      - OR REFACTOR TO INCLUDE AS MATRIX SPEC:
+            #          IF MATRIX IS 1D, USE AS OUTPUT TEMPLATE
+            #          IF ALL ITS VALUES ARE 1'S => FULL CONNECTIVITY MATRIX
+            #          IF ALL ITS VALUES ARE 0'S => RANDOM CONNECTIVITY MATRIX
+            #          NOTE:  NO NEED FOR IDENTITY MATRIX, AS THAT WOULD BE SQUARE SO NO NEED FOR OUTPUT TEMPLATE
+            #      - DOCUMENT WHEN DONE
+            # MODIFIED 3/26/17 OLD:
+            # Check for and validate kwReceiver first, since it may be needed to validate and/or construct the matrix
+            # First try to get receiver from specification in params
+            if RECEIVER in param_set:
+                self.receiver = param_set[RECEIVER]
+                # Check that specification is a list of numbers or an array
+                if ((isinstance(self.receiver, list) and all(
+                        isinstance(elem, numbers.Number) for elem in self.receiver)) or
+                        isinstance(self.receiver, np.ndarray)):
+                    self.receiver = np.atleast_1d(self.receiver)
+                else:
+                    raise FunctionError("receiver param ({0}) for {1} must be a list of numbers or an np.array".
+                                        format(self.receiver, self.name))
+            # No receiver, so use sender as template (assuming square -- e.g., identity -- matrix)
+            else:
+                if (self.owner and self.owner.prefs.verbosePref) or self.prefs.verbosePref:
+                    print("Identity matrix requested but kwReceiver not specified; sender length ({0}) will be used".
+                          format(sender_len))
+                self.receiver = param_set[RECEIVER] = sender
+
+            receiver_len = len(self.receiver)
+
+            # Check rest of params
+            message = ""
+            for param_name, param_value in param_set.items():
+
+                # Receiver param already checked above
+                if param_name == RECEIVER:
+                    continue
+
+                # Not currently used here
+                if param_name in function_keywords:
+                    continue
+
+                if param_name == HAS_INITIALIZERS:
+                    continue
+
+                # Matrix specification param
+                elif param_name == MATRIX:
+
+                    # A number (to be used as a filler), so OK
+                    if isinstance(param_value, numbers.Number):
+                        continue
+
+                    # np.matrix or np.ndarray provided, so validate that it is numeric and check dimensions
+                    elif isinstance(param_value, (list, np.ndarray, np.matrix)):
+                        # get dimensions specified by:
+                        #   variable (sender): width/cols/outer index
+                        #   kwReceiver param: height/rows/inner index
+
+                        weight_matrix = np.atleast_2d(param_value)
+                        if 'U' in repr(weight_matrix.dtype):
+                            raise FunctionError("Non-numeric entry in MATRIX "
+                                                "specification ({}) for the {} "
+                                                "function of {}".format(param_value,
+                                                                        self.name,
+                                                                        self.owner_name))
+
+                        if weight_matrix.ndim != 2:
+                            raise FunctionError("The matrix provided for the {} function of {} must be 2d (it is {}d".
+                                                format(weight_matrix.ndim, self.name, self.owner_name))
+
+                        matrix_rows = weight_matrix.shape[0]
+                        matrix_cols = weight_matrix.shape[1]
+
+                        # Check that number of rows equals length of sender vector (variable)
+                        if matrix_rows != sender_len:
+                            raise FunctionError("The number of rows ({}) of the "
+                                                "matrix provided for {} function "
+                                                "of {} does not equal the length "
+                                                "({}) of the sender vector "
+                                                "(variable)".format(matrix_rows,
+                                                                    self.name,
+                                                                    self.owner_name,
+                                                                    sender_len))
+
+                    # Auto, full or random connectivity matrix requested (using keyword):
+                    # Note:  assume that these will be properly processed by caller
+                    #        (e.g., MappingProjection._instantiate_receiver)
+                    elif is_matrix_keyword(param_value):
+                        continue
+
+                    # Identity matrix requested (using keyword), so check send_len == receiver_len
+                    elif param_value in {IDENTITY_MATRIX, HOLLOW_MATRIX}:
+                        # Receiver length doesn't equal sender length
+                        if not (self.receiver.shape == sender.shape and self.receiver.size == sender.size):
+                            # if self.owner.prefs.verbosePref:
+                            #     print ("Identity matrix requested, but length of receiver ({0})"
+                            #            " does not match length of sender ({1});  sender length will be used".
+                            #            format(receiver_len, sender_len))
+                            # # Set receiver to sender
+                            # param_set[kwReceiver] = sender
+                            raise FunctionError("{} requested for the {} function of {}, "
+                                                "but length of receiver ({}) does not match length of sender ({})".
+                                                format(param_value, self.name, self.owner_name, receiver_len,
+                                                       sender_len))
+                        continue
+
+                    # list used to describe matrix, so convert to 2D array and pass to validation of matrix below
+                    elif isinstance(param_value, list):
+                        try:
+                            param_value = np.atleast_2d(param_value)
+                        except (ValueError, TypeError) as error_msg:
+                            raise FunctionError(
+                                "Error in list specification ({}) of matrix for the {} function of {}: {})".
+                                    # format(param_value, self.__class__.__name__, error_msg))
+                                    format(param_value, self.name, self.owner_name, error_msg))
+
+                    # string used to describe matrix, so convert to np.array and pass to validation of matrix below
+                    elif isinstance(param_value, str):
+                        try:
+                            param_value = np.atleast_2d(param_value)
+                        except (ValueError, TypeError) as error_msg:
+                            raise FunctionError("Error in string specification ({}) of the matrix "
+                                                "for the {} function of {}: {})".
+                                                # format(param_value, self.__class__.__name__, error_msg))
+                                                format(param_value, self.name, self.owner_name, error_msg))
+
+                    # function so:
+                    # - assume it uses random.rand()
+                    # - call with two args as place markers for cols and rows
+                    # -  validate that it returns an array
+                    elif isinstance(param_value, types.FunctionType):
+                        test = param_value(1, 1)
+                        if not isinstance(test, np.ndarray):
+                            raise FunctionError("A function is specified for the matrix of the {} function of {}: {}) "
+                                                "that returns a value ({}) that is not an array".
+                                                # format(param_value, self.__class__.__name__, test))
+                                                format(self.name, self.owner_name, param_value, test))
+
+                    elif param_value is None:
+                        raise FunctionError("TEMP ERROR: param value is None.")
+
+                    else:
+                        raise FunctionError("Value of {} param ({}) for the {} function of {} "
+                                            "must be a matrix, a number (for filler), or a matrix keyword ({})".
+                                            format(param_name,
+                                                   param_value,
+                                                   self.name,
+                                                   self.owner_name,
+                                                   MATRIX_KEYWORD_NAMES))
+                else:
+                    continue
+            if message:
+                raise FunctionError(message)
+
+        # owner is a mechanism, state
+        # OR function was defined on its own (no owner)
+        else:
+            if MATRIX in param_set:
+                param_value = param_set[MATRIX]
+
+                # numeric value specified; verify that it is compatible with variable
+                if isinstance(param_value, (float, list, np.ndarray, np.matrix)):
+                    param_size = np.size(np.atleast_2d(param_value), 0)
+                    param_shape = np.shape(np.atleast_2d(param_value))
+                    variable_size = np.size(np.atleast_2d(self.defaults.variable),1)
+                    variable_shape = np.shape(np.atleast_2d(self.defaults.variable))
+                    if param_size != variable_size:
+                        raise FunctionError("Specification of matrix and/or default_variable for {} is not valid. The "
+                                            "shapes of variable {} and matrix {} are not compatible for multiplication".
+                                            format(self.name, variable_shape, param_shape))
+
+                # keyword matrix specified - not valid outside of a projection
+                elif is_matrix_keyword(param_value):
+                    raise FunctionError("{} is not a valid specification for the matrix parameter of {}. Keywords "
+                                        "may only be used to specify the matrix parameter of a Projection's "
+                                        "LinearTransform function. When the LinearTransform function is implemented in a "
+                                        "mechanism, such as {}, the correct matrix cannot be determined from a "
+                                        "keyword. Instead, the matrix must be fully specified as a float, list, "
+                                        "np.ndarray".
+                                        format(param_value, self.name, self.owner.name))
+
+                # The only remaining valid option is matrix = None (sorted out in instantiate_attribs_before_fn)
+                elif param_value is not None:
+                    raise FunctionError("Value of the matrix param ({}) for the {} function of {} "
+                                        "must be a matrix, a number (for filler), or a matrix keyword ({})".
+                                        format(param_value,
+                                               self.name,
+                                               self.owner_name,
+                                               MATRIX_KEYWORD_NAMES))
+
+    def _instantiate_attributes_before_function(self, function=None, context=None):
+        # replicates setting of receiver in _validate_params
+        if isinstance(self.owner, Projection):
+            self.receiver = copy_parameter_value(self.defaults.variable)
+
+        matrix = self.parameters.matrix._get(context)
+
+        if matrix is None and not hasattr(self.owner, "receiver"):
+            variable_length = np.size(np.atleast_2d(self.defaults.variable), 1)
+            matrix = np.identity(variable_length)
+        self.parameters.matrix._set(self.instantiate_matrix(matrix), context)
+
+    def instantiate_matrix(self, specification, context=None):
+        """Implements matrix indicated by specification
+
+         Specification is derived from MATRIX param (passed to self.__init__ or self._function)
+
+         Specification (validated in _validate_params):
+            + single number (used to fill self.matrix)
+            + matrix keyword (see get_matrix)
+            + 2D list or np.ndarray of numbers
+
+        :return matrix: (2D list)
+        """
+        from psyneulink.core.components.projections.projection import Projection
+        if isinstance(self.owner, Projection):
+            # Matrix provided (and validated in _validate_params); convert to array
+            if isinstance(specification, np.matrix):
+                return np.array(specification)
+
+            sender = copy_parameter_value(self.defaults.variable)
+            sender_len = sender.shape[0]
+            try:
+                receiver = self.receiver
+            except:
+                raise FunctionError("Can't instantiate matrix specification ({}) for the {} function of {} "
+                                    "since its receiver has not been specified".
+                                    format(specification, self.name, self.owner_name))
+                # receiver = sender
+            receiver_len = receiver.shape[0]
+
+            matrix = get_matrix(specification, rows=sender_len, cols=receiver_len, context=context)
+
+            # This should never happen (should have been picked up in validate_param or above)
+            if matrix is None:
+                raise FunctionError("MATRIX param ({}) for the {} function of {} must be a matrix, a function "
+                                    "that returns one, a matrix specification keyword ({}), or a number (filler)".
+                                    format(specification, self.name, self.owner_name, MATRIX_KEYWORD_NAMES))
+            else:
+                return matrix
+        else:
+            return np.array(specification)
+
+
+    def _gen_llvm_function_body(self, ctx, builder, params, state, arg_in, arg_out, *, tags:frozenset):
+        # Restrict to 1d arrays
+        if self.defaults.variable.ndim != 1:
+            warnings.warn("Shape mismatch: {} (in {}) got 2D input: {}".format(
+                          self, self.owner, self.defaults.variable),
+                          pnlvm.PNLCompilerWarning)
+            arg_in = builder.gep(arg_in, [ctx.int32_ty(0), ctx.int32_ty(0)])
+        if self.defaults.value.ndim != 1:
+            warnings.warn("Shape mismatch: {} (in {}) has 2D output: {}".format(
+                          self, self.owner, self.defaults.value),
+                          pnlvm.PNLCompilerWarning)
+            arg_out = builder.gep(arg_out, [ctx.int32_ty(0), ctx.int32_ty(0)])
+
+        matrix = ctx.get_param_or_state_ptr(builder, self, MATRIX, param_struct_ptr=params, state_struct_ptr=state)
+        normalize = ctx.get_param_or_state_ptr(builder, self, NORMALIZE, param_struct_ptr=params)
+
+        # Convert array pointer to pointer to the fist element
+        matrix = builder.gep(matrix, [ctx.int32_ty(0), ctx.int32_ty(0)])
+        vec_in = builder.gep(arg_in, [ctx.int32_ty(0), ctx.int32_ty(0)])
+        vec_out = builder.gep(arg_out, [ctx.int32_ty(0), ctx.int32_ty(0)])
+
+        input_length = ctx.int32_ty(arg_in.type.pointee.count)
+        output_length = ctx.int32_ty(arg_out.type.pointee.count)
+
+        # if normalize:
+        #     if vec_in is not zeros:
+        #     # FIX: NORMALIZE vec_in and matrix here
+        #         vec_in_sum = fsum(builder, vec_in)
+        #         vec_in = fdiv(builder, vec_in, vec_in_sum)
+        #     if matrix is not zeros:
+        #     # FIX: NORMALIZE matrix here
+
+        builtin = ctx.import_llvm_function("__pnl_builtin_vxm")
+        builder.call(builtin, [vec_in, matrix, input_length, output_length, vec_out])
+        return builder
+
+    def _function(self,
+                 variable=None,
+                 context=None,
+                 params=None,
+                 ):
+        """
+
+        Arguments
+        ---------
+        variable : list or 1d array
+            array to be transformed;  length must equal the number of rows of `matrix <LinearTransform.matrix>`.
+
+        params : Dict[param keyword: param value] : default None
+            a `parameter dictionary <ParameterPort_Specification>` that specifies the parameters for the
+            function.  Values specified for parameters in the dictionary override any assigned to those parameters in
+            arguments of the constructor.
+
+        Returns
+        ---------
+
+        dot product of variable and matrix : 1d array
+            length of the array returned equals the number of columns of `matrix <LinearTransform.matrix>`.
+
+        """
+        vector = np.array(variable)
+        matrix = self._get_current_parameter_value(MATRIX, context)
+        normalize = self._get_current_parameter_value(NORMALIZE, context)
+        if normalize:
+            if np.any(vector):
+                vector = vector / np.linalg.norm(vector)
+            if np.any(matrix):
+                # FIX: the axis along which norming is carried out should probably be a parameter
+                #      Also need to deal with column- (or row-) wise zeros which cause div by zero
+                #      Replace columns (if norming axis 0) or rows (if norming axis 1) of zeros with 1's
+                # matrix = matrix / np.linalg.norm(matrix,axis=-1,keepdims=True)
+                matrix = matrix / np.linalg.norm(matrix,axis=0,keepdims=True)
+
+        result = np.dot(vector, matrix)
+        return self.convert_output_type(result)
+
+    @staticmethod
+    def keyword(obj, keyword):
+
+        from psyneulink.core.components.projections.pathway.mappingprojection import MappingProjection
+        rows = None
+        cols = None
+        # use of variable attribute here should be ok because it's using it as a format/type
+        if isinstance(obj, MappingProjection):
+            if isinstance(obj.sender.defaults.value, numbers.Number):
+                rows = 1
+            else:
+                rows = len(obj.sender.defaults.value)
+            if isinstance(obj.receiver.defaults.variable, numbers.Number):
+                cols = 1
+            else:
+                cols = obj.receiver.socket_width
+        matrix = get_matrix(keyword, rows, cols)
+
+        if matrix is None:
+            raise FunctionError("Unrecognized keyword ({}) specified for the {} function of {}".
+                                format(keyword, obj.name, obj.owner_name))
+        else:
+            return matrix
+
+    def param_function(owner, function):
+        sender_len = len(owner.sender.defaults.value)
+        receiver_len = len(owner.receiver.defaults.variable)
+        return function(sender_len, receiver_len)
+
+    def _is_identity(self, context=None, defaults=False):
+        if defaults:
+            matrix = self.defaults.matrix
+        else:
+            matrix = self.parameters.matrix._get(context)
+
+        # if matrix is not an np array with at least one dimension,
+        # this isn't an identity matrix
+        try:
+            size = matrix.shape[0]
+        except (AttributeError, IndexError):
+            return False
+
+        # check if the matrix is the same as the identity matrix
+        # note that we can use the first dimension size to create the identity matrix
+        # because if the matrix is not square, this comparison will fail anyway
+        identity_matrix = np.identity(size)
+        # numpy has deprecated == comparisons of arrays
+        try:
+            return np.array_equal(matrix, identity_matrix)
+        except TypeError:
+            return matrix == identity_matrix
+
+# def is_matrix_spec(m):
+#     if m is None:
+#         return True
+#     if isinstance(m, (list, np.ndarray, types.FunctionType)):
+#         return True
+#     if m in MATRIX_KEYWORD_VALUES:
+#         return True
+#     return False
+
 
 
 class CombineMeans(CombinationFunction):  # ------------------------------------------------------------------------
