+++ psyneulink/core/globals/keywords.py
@@ -76,7 +76,7 @@ __all__ = [
     'LEARNING_MECHANISMS', 'LEARNING_PATHWAY', 'LEARNING_PROJECTION', 'LEARNING_PROJECTION_PARAMS', 'LEARNING_RATE',
     'LEARNING_SCALE', 'LEARNING_SCALE_LITERALS', 'LEARNING_SCALE_NAMES', 'LEARNING_SIGNAL', 'LEARNING_SIGNAL_SPECS',
     'LEARNING_SIGNALS', 'LESS_THAN', 'LESS_THAN_OR_EQUAL', 'LINEAR', 'LINEAR_COMBINATION_FUNCTION', 'LINEAR_FUNCTION',
-    'LINEAR_MATRIX_FUNCTION', 'LOG_ENTRIES', 'LOGISTIC_FUNCTION', 'Loss', 'LOSSES', 'LOW', 'LVOC_CONTROL_MECHANISM',
+    'LINEAR_TRANSFORM_FUNCTION', 'LOG_ENTRIES', 'LOGISTIC_FUNCTION', 'Loss', 'LOSSES', 'LOW', 'LVOC_CONTROL_MECHANISM',
     'MAPPING_PROJECTION', 'MAPPING_PROJECTION_PARAMS', 'MASKED_MAPPING_PROJECTION',
     'MATRIX', 'MATRIX_KEYWORD_NAMES', 'MATRIX_KEYWORD_SET', 'MATRIX_KEYWORD_VALUES', 'MATRIX_KEYWORDS','MatrixKeywords',
     'MATRIX_WEIGHTS', 'MAX_ABS_DIFF', 'MAX_ABS_INDICATOR', 'MAX_ONE_HOT', 'MAX_ABS_ONE_HOT', 'MAX_ABS_VAL',
@@ -745,6 +745,7 @@ REDUCE_FUNCTION = "Reduce Function"
 CONCATENATE_FUNCTION = "Concatenate Function"
 REARRANGE_FUNCTION = 'Rearrange Function'
 LINEAR_COMBINATION_FUNCTION = "LinearCombination Function"
+LINEAR_TRANSFORM_FUNCTION = "LinearTransform Function"
 COMBINE_MEANS_FUNCTION = "CombineMeans Function"
 
 # TransferFunctions:
@@ -761,7 +762,6 @@ GAUSSIAN_DISTORT_FUNCTION = "GaussianDistort Function"
 BINOMIAL_DISTORT_FUNCTION = 'BinomialDistort Function'
 DROPOUT_FUNCTION = 'Dropout Function'
 SOFTMAX_FUNCTION = 'SoftMax Function'
-LINEAR_MATRIX_FUNCTION = "LinearTransform Function"
 TRANSFER_WITH_COSTS_FUNCTION = "TransferWithCosts Function"
 
 # SelectionFunctions:
