+++ psyneulink/library/components/projections/pathway/autoassociativeprojection.py
@@ -105,7 +105,7 @@ from beartype import beartype
 from psyneulink._typing import Optional
 
 from psyneulink.core.components.component import parameter_keywords
-from psyneulink.core.components.functions.nonstateful.transferfunctions import LinearTransform
+from psyneulink.core.components.functions.nonstateful.combinationfunctions import LinearTransform
 from psyneulink.core.components.functions.function import get_matrix
 from psyneulink.core.components.projections.pathway.mappingprojection import MappingError, MappingProjection
 from psyneulink.library.components.projections.pathway.maskedmappingprojection import MaskedMappingProjection
